Added a gitignore updated README with bage Added deploy.sh Updated check and build for coding style

+++ build.sh
@@ -1,22 +1,25 @@
 #!/bin/bash -e
 . /etc/profile.d/modules.sh
 SOURCE_FILE=${NAME}-${VERSION}.tar.gz
-echo ${SRC_DIR}
 module load ci
-# SRC_DIR=/repo/src/$NAME
-module add gcc/4.8.4
-
-# GSL can use FFTW and ATLAS
-module add atlas
-module add fftw/3.3.4
+mkdir -p ${SRC_DIR}
 echo "getting the file from gnu.org mirror"
-if [[ ! -s ${SRC_DIR}/${SOURCE_FILE} ]] ; then
-  mkdir -vp ${SRC_DIR}
-#wget --verbose ftp://ftp.is.co.za/mirror/ftp.gnu.org/gnu/gsl/$SOURCE_FILE -O $SRC_DIR/$SOURCE_FILE
+
+if [ ! -e ${SRC_DIR}/${SOURCE_FILE}.lock ] && [ ! -s ${SRC_DIR}/${SOURCE_FILE} ] ; then
+# claim the download
+  touch  ${SRC_DIR}/${SOURCE_FILE}.lock
   wget http://mirror.ufs.ac.za/gnu/gnu/${NAME}/${NAME}-${VERSION}.tar.gz -O ${SRC_DIR}/${SOURCE_FILE}
+  echo "releasing lock"
+  rm -v ${SRC_DIR}/${SOURCE_FILE}.lock
+elif [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; then
+  # Someone else has the file, wait till it's released
+  while [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; do
+    echo " There seems to be a download currently under way, will check again in 5 sec"
+    sleep 5
+  done
 fi
 ls -lht ${SRC_DIR}/${SOURCE_FILE}
-ctar xfz ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE}
+tar xfz ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE}
 cd ${WORKSPACE}/${NAME}-${VERSION}
 ./configure --prefix ${SOFT_DIR}
 make
